requirements.txt
Flask==3.0.3
google-generativeai==0.7.2
PyPDF2==3.0.1
python-docx==1.1.2